backend/requirements.txt
aerich
asyncpg
fastapi
tortoise-orm
uvicorn